fix the script in background

+++ scripts/klaytnScriptUnderground.ts
@@ -17,7 +17,7 @@ const NETWORKNAME:networkNameType = "klaytn";
  */
 function readFilesSync(dir: any) {
   const files: any[] = [];
-  console.log(dir);
+
   fs.readdirSync(dir).forEach((filename) => {
     const name = path.parse(filename).name;
     const ext = path.parse(filename).ext;
@@ -52,10 +52,14 @@ async function main() {
   const signer = getSigner(NETWORKNAME);
   const account = signer.address;
   while(1) {
-    console.log("_____new interval klaytn___")
+    console.log("_____new interval polygon___")
     const nextNonce = await provider.getTransactionCount(account, "latest");
+    const fileDir = path.join(__dirname, "../transaction_log/new_contract/"+NETWORKNAME+"/unsynced/");
+    if(!fs.existsSync(fileDir)) {
+      continue;
+    }
     const files = readFilesSync(
-      path.join(__dirname, "../../transaction_log/new_contract/"+NETWORKNAME+"/unsynced/")
+      path.join(fileDir)
     );
     for (let i = 0; i < files.length; i++) {
       const thisNonce = nextNonce + i;
@@ -84,17 +88,23 @@ async function main() {
         file["moneyDetails"] = moneyDetails;
         file["networkName"] = NETWORKNAME;
         file["functionName"] = functionName;
+        
+        
       } catch (e:any) {
         // ignore
         console.log("NO MONEY ????");
         return;
       }
       // save file 
-      const filePath = path.join(__dirname, "../../transaction_log/new_contract/"+NETWORKNAME+"/synced/" + fileName + ".json");
+      const filePath = path.join(__dirname, "../transaction_log/new_contract/"+NETWORKNAME+"/synced/" + fileName + ".json");
+      if (!fs.existsSync(filePath)) 
+        continue;
       fs.writeFileSync(filePath, JSON.stringify(file));
 
       // delete the old file
-      const oldFilePath = path.join(__dirname, "../../transaction_log/new_contract/"+NETWORKNAME+"/unsynced/" + fileName + ".json");
+      const oldFilePath = path.join(__dirname, "../transaction_log/new_contract/"+NETWORKNAME+"/unsynced/" + fileName + ".json");
+      if (!fs.existsSync(oldFilePath)) 
+        continue;
       fs.unlinkSync(oldFilePath);
     }
     await delay(2000);
